+++ examples/estimate-swap/CHANGELOG.md
@@ -1,5 +1,24 @@
 # xchainjs-estimate-swap
 
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [7d70a1c]
+- Updated dependencies [7d70a1c]
+- Updated dependencies [7d70a1c]
+  - @xchainjs/xchain-midgard@0.5.8
+  - @xchainjs/xchain-thornode@0.3.17
+  - @xchainjs/xchain-midgard-query@0.1.20
+  - @xchainjs/xchain-thorchain-query@0.7.17
+
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [7df3870]
+  - @xchainjs/xchain-thorchain-query@0.7.16
+
 ## 1.0.8
 
 ### Patch Changes